DX-1637: Visually appalling as the design isn't in place with the boxes.
The text misses an example that I guess will need a screenshot.

The main index page went completely FUBAR when I tried goofing around with the boxes. I left it unchanged to not send everything down the toilet.

Merchant managed is missing it's set up page as it isn't included in this branch. Will add when MM is approved and merged.