### Version 3.1.0
Summer has finally arrived, and we have quite a treat waiting in our final
release before the vacation pulse kicks in. Without further ado: We now proudly
offer **Apple Pay**! It is available in all our Checkout v3 implementations and
integrations. Give it a go in our [demoshop][demoshop]! There are a couple of
other new additions as well:

*   Our [SDKs][sdk-modules] have been updated to support Checkout v3
*   You can now [request delivery information][request-delivery-information] for
  selected instruments
*   An eligibility check [eligibility check][eligibility-check] for wallets has
  been added to instrument mode
*   "Remember me" has been added to Checkout v3 [Starter][checkout-v3-starter]
  and [Business][checkout-v3-business]
*   Fixed typos, minor bugs and code examples