DX-2004: Fix links in migration guide
Add a custom `CLinkBlock` Liquid tag which only renders a link if the
link is absolute, or (when it is relative) is an existing local `.md`
file. This makes it possible to use the same field includes in tables
outside of the regular documentation sections, such as in the migration
guide which resides in the `checkout-v3` root directory.

Also use more of the includes elsewhere to test the new `clink` usage
to see that it works.

Refactor the `status` field description so it is much shorter and move
the longer status description to a separate `status.md` include that
is again included in a new `status.md` in all technical references.